+++ src/types.ts
@@ -219,10 +219,7 @@ export type StorableWithTimestamp = {
     value: StorableObj;
 };
 
-export interface TagWithUsageCounts extends mastodon.v1.Tag {
-    numAccounts?: number;
-    numToots?: number;
-};
+export interface TagWithUsageCounts extends mastodon.v1.Tag, TrendingHistory {};
 
 // Similar to mastodon.v1.Context: https://docs.joinmastodon.org/entities/Context/
 export type TootContext = {
@@ -241,11 +238,7 @@ export type TootScore = {
 };
 
 export type TootScores = Record<ScoreName, WeightedScore>;
-
-export interface TrendingLink extends mastodon.v1.TrendLink {
-    numToots?: number;
-    numAccounts?: number;
-};
+export interface TrendingLink extends mastodon.v1.TrendLink, TrendingHistory {};
 
 export interface TrendingStorage {
     links: TrendingLink[];
@@ -255,6 +248,12 @@ export interface TrendingStorage {
 
 export type TrendingWithHistory = TagWithUsageCounts | TrendingLink;
 export type TrendingObj = TrendingWithHistory | Toot;
+
+export interface TrendingHistory {
+    numAccounts?: number;
+    numToots?: number;
+    regex?: RegExp;
+};
 
 export type WeightedScore = {
     raw: number;